fix: apply Karpathy Guidelines review results (#41)
* fix: apply Karpathy Guidelines review results

- Fix HIGH: spurious sensors_tick_50ms/100ms outside timer gate
- Fix HIGH: ETB PID control now runs at enforced 2ms cadence
- Fix MED: add static_assert + documentation for kTriggerTooth0EngineDeg
- Fix MED: remove unused dlc parameter from can_filters_should_accept()
- Fix MED: consolidate avg4/avg8 into avg_n(), iir_alpha_03/01 into iir()
- Fix LOW: remove dead code system_reset() and ckp_angle_to_ticks()
- Fix LOW: remove duplicate gpio_set_af() out-of-line implementation

All changes verified with 'make host-test' (MVP tests pass)

* fix: apply Karpathy Guidelines review

---------

+++ src/app/can_stack.h
@@ -22,7 +22,7 @@ void can_stack_process(uint32_t now_ms,
                        int8_t  stft_pct,
                        uint8_t vvt_intake_pct,
                        uint8_t vvt_exhaust_pct,
-                       uint8_t status_bits) noexcept;
+                       uint16_t status_bits) noexcept;
 
 // Lambda × 1000 do último frame WBO2 recebido (raw, sem fallback)
 uint16_t can_stack_lambda_milli() noexcept;
